Update DatePickerX.js
Added altFormat and altInput to options. You can use these to add the selected date to a second form input. Set altFormat to the date format required and altInput to the ID of the form input. For example you can set 'format' to the format you want the user to see and set 'altFormat' to something more convenient for your back-end program